Split appearance and permitted source directories
