+++ src/main/resources/io/jenkins/plugins/prism/PrismAppearanceConfiguration/config.jelly
@@ -1,0 +1,10 @@
+<?jelly escape-by-default='true'?>
+<j:jelly xmlns:j="jelly:core" xmlns:f="/lib/form">
+
+  <f:section title="${%Prism syntax highlighting}">
+    <f:entry title="${%Theme}" field="theme">
+      <f:select default="PRISM"/>
+    </f:entry>
+  </f:section>
+
+</j:jelly>
