+++ src/main/resources/io/jenkins/plugins/prism/PrismAppearanceConfiguration/config.properties
@@ -1,0 +1,2 @@
+sourceDirectories.title=Permitted Source Code Directories
+sourceDirectories.description=Valid and permitted source code locations on agents (outside the workspace).
