+++ src/main/resources/io/jenkins/plugins/prism/PrismAppearanceConfiguration/help-sourceDirectories.html
@@ -1,0 +1,7 @@
+<div>
+    Some plugins copy source code files to Jenkins' build folder so that these files can be rendered in the user
+    interface together with build results (coverage, warnings, etc.). If these files are not part of the workspace
+    of a build then Jenkins will not show them by default: otherwise sensitive files could be shown by accident.
+    You can provide a list of additional source code directories that are allowed to be shown in Jenkins user interface
+    here. Note, that such a directory must be an absolute path on the <b>agent</b> that executes the build.
+</div>
